Try to handle dependency issues in Rails better.
Rails annoyingly loads all helpers for any subclass of ActionController::Base.
This includes the TestController used by ActionView::TestCase. So basically
our helpers are loaded just by loading Rails test stuff.

Our solution is to use 'load' instead of 'require'. This means that some
files will load more than once but it means we can capture the tests even
though Rails has already loaded our helpers. We already had code in place
to prevent the test from registering more than once.

The only downside is a file may not be designed to be reloadable (i.e.
it might set some constant so we will start getting warnings, etc). I think
we are just going to have to cross that bridge when we come to it. With the
way things were the helper tests were not getting run at all.