Removed name randomization.
When generating test names (required by Test::Unit) we used to generate a
random name (if one not provided). In case of clashes (common if providing
a name) it also added a 5 digit number to the end of every name.

Later we found it useful to include the file and line number of the test into
the name. This made the randomization and the 5 digit number unnecessary.
This commit removes that making the output much cleaner and reducing the amount
of code in this library.