Better Rails integration.
The backtrace was getting overly stripped because rails assumed all tests
are in the "test" directoy. This fixes this by removing the backtrace silencer
when in the test environment. We might consider allowing silencing of backtrace
in libraries to make it more like Rails' default behavior. But I have always
found that annoying. We are debugging when we are running tests so having the
full backtrace is necessary.

Also fork running the tests so we are sure it is happening in isolation.